+++ src/tool/99-main.sh
@@ -10,8 +10,14 @@ main() {
 
   case "$ACTION" in
     onboard)     onboard ;;
-    config)      configure ;;
-    show-config) show_config ;;
+    config)
+      if [[ "$GC_HC_CONFIG_SHOW" == "true" ]]; then show_all_config
+      elif [[ "$GC_HC_CONFIG_SMTP" == "true" && -n "$GC_HC_MAIL_TEST" ]]; then send_test_mail_saved "$GC_HC_MAIL_TEST"
+      elif [[ "$GC_HC_CONFIG_SMTP" == "true" && "$YES" == "true" ]]; then configure_mail_noninteractive
+      elif [[ "$GC_HC_CONFIG_SMTP" == "true" ]]; then configure_mail
+      else configure
+      fi
+      ;;
     check)       run_check ;;
     status)      show_status ;;
     logs)        show_logs ;;